requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
Django==4.0.2
django-appconf==1.0.5
django-cryptography==1.0
django-decouple==2.1
django-ranged-response==0.2.0
django-widget-tweaks==1.4.12
idna==3.3
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
